refactor: changed name and images for Zorkcoin
Much of this was done via scripts so there may still be
fine tuning to be done in the future. There shouldn't be
changes other than additions to existing copyright claims.

Manual modifications to many of the language files was made
and these will likely need corrected via the normal translation
process in the future.

Resolves: #2